pyproject.toml
[project]
name = "dio-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Gustavo Cassiano",email = "gucpinto26@gmail.com"}
]

requires-python = ">=3.14"
dependencies = [
    "fastapi",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "databases[aiosqlite,asyncpg]",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2-binary",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic",
    "email-validator (>=2.0.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
]

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
pytest-asyncio = ">=1.3.0,<2.0.0"
pytest-mock = ">=3.15.1,<4.0.0"
httpx = ">=0.28.1,<0.29.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false